code refactoring and service implementation is added
