+++ src/main/java/com/personal/TravelPlanner/service/AuthenticationService.java
@@ -1,118 +1,17 @@
 package com.personal.TravelPlanner.service;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.personal.TravelPlanner.dto.AuthenticationRequestDTO;
 import com.personal.TravelPlanner.dto.AuthenticationResponseDTO;
 import com.personal.TravelPlanner.dto.RegisterRequestDTO;
-import com.personal.TravelPlanner.entity.Token;
-import com.personal.TravelPlanner.entity.User;
-import com.personal.TravelPlanner.exception.TokenExpiredException;
-import com.personal.TravelPlanner.repository.TokenRepository;
-import com.personal.TravelPlanner.repository.UserRepository;
-import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
 
-import java.io.IOException;
+public interface AuthenticationService {
 
-@Service
-@RequiredArgsConstructor
-public class AuthenticationService  {
-
-
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-    private final JwtService jwtService;
-    private final AuthenticationManager authenticationManager;
-    private final TokenRepository tokenRepository;
-
-    public AuthenticationResponseDTO register(RegisterRequestDTO request) {
-        var user = User.builder()
-                .firstname(request.getFirstname())
-                .lastname(request.getLastname())
-                .email(request.getEmail())
-                .password(passwordEncoder.encode(request.getPassword()))
-                .role(request.getRole())
-                .build();
-        var savedUser = userRepository.save(user);
-
-        var jwtToken = jwtService.generateToken(user);
-        var refreshToken = jwtService.generateRefreshToken(user);
-        saveUserToken(savedUser, jwtToken);
-        return AuthenticationResponseDTO.builder()
-                .accessToken(jwtToken).refreshToken(refreshToken)
-                .build();
-    }
-
-    public AuthenticationResponseDTO authenticate(AuthenticationRequestDTO request) {
-        authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(
-                        request.getEmail(),
-                        request.getPassword()
-                )
-        );
-        var user = userRepository.findByEmail(request.getEmail()).orElseThrow();
-        var jwtToken= jwtService.generateToken(user);
-        var refreshToken = jwtService.generateRefreshToken(user);
-        saveUserToken(user, jwtToken);
-        return AuthenticationResponseDTO.builder()
-                .accessToken(jwtToken).refreshToken(refreshToken)
-                .build();
-    }
-
-    public void refreshToken (
+    AuthenticationResponseDTO register(RegisterRequestDTO request);
+    AuthenticationResponseDTO authenticate(AuthenticationRequestDTO request);
+     void refreshToken (
             HttpServletRequest request,
             HttpServletResponse response
-    ) throws Exception {
-        final String authHeader = request.getHeader(HttpHeaders.AUTHORIZATION);
-        final String refreshToken;
-        final String userEmail;
-        if (authHeader == null ||!authHeader.startsWith("Bearer ")) {
-            return;
-        }
-        refreshToken = authHeader.substring(7);
-        userEmail = jwtService.extractUsername(refreshToken);
-        if (userEmail != null) {
-            var user = this.userRepository.findByEmail(userEmail)
-                    .orElseThrow();
-            if (jwtService.isTokenValid(refreshToken, user)) {
-                var accessToken = jwtService.generateToken(user);
-                revokeAllUserTokens(user);
-                saveUserToken(user, accessToken);
-                var authResponse = AuthenticationResponseDTO.builder()
-                        .accessToken(accessToken)
-                        .build();
-                new ObjectMapper().writeValue(response.getOutputStream(), authResponse);
-            }
-        }
-    }
-
-
-    private void saveUserToken(User user, String jwtToken) {
-        var token = Token.builder()
-                .user(user)
-                .token(jwtToken)
-                .expired(false)
-                .revoked(false)
-                .build();
-        tokenRepository.save(token);
-    }
-
-    private void revokeAllUserTokens(User user) {
-        var validUserTokens = tokenRepository.findByUser_IdAndExpiredIsFalseAndRevokedIsFalse(user.getId());
-        if (validUserTokens.isEmpty())
-            return;
-        validUserTokens.forEach(token -> {
-            token.setExpired(true);
-            token.setRevoked(true);
-        });
-        tokenRepository.saveAll(validUserTokens);
-    }
-
+    ) throws Exception;
 }
